requirements.txt
aiohttp==3.7.4.post0
aioredis==1.3.1
aioredis-lock==0.1.0
alexflipnote.py==2.3.0
appdirs==1.4.4
astunparse==1.6.3
async-timeout==3.0.1
asyncpg==0.22.0
attrs==20.3.0
braceexpand==0.1.6
certifi==2020.12.5
cfgv==3.2.0
chardet==4.0.0
click==7.1.2
colorama==0.4.4
colorthief==0.2.1
discord-ext-ipc==1.1.2a0
discord-py-slash-command==1.0.9.5
discord.py==1.6.0
disputils==0.1.3
distlib==0.3.1
dump==0.0.5
env==0.1.0
filelock==3.0.12
hiredis==1.1.0
humanize==3.2.0
hurry==1.1
hurry.filesize==0.9
identify==2.1.2
idna==2.10
imgurpython==1.1.7
import-expression==1.1.4
jishaku==1.20.0.220
lxml==4.6.2
marshmallow==3.10.0
motor==2.3.1
multidict==5.1.0
mystbin.py==1.0.1
nodeenv==1.5.0
numpy==1.20.1
packaging==20.9
parsedatetime==2.6
Pillow==8.1.2
pluggy==0.13.1
pre-commit==2.11.1
psutil==5.8.0
py==1.10.0
pymongo==3.11.3
pyparsing==2.4.7
python-dateutil==2.8.1
PyYAML==5.4.1
requests==2.25.1
six==1.15.0
speedtest-cli==2.1.2
suntime==1.2.5
toml==0.10.2
tox==3.23.0
translate==3.5.0
typing-extensions==3.7.4.3
umongo==3.0.0b10
urllib3==1.26.3
virtualenv==20.4.2
virtualenv-clone==0.5.4
yarl==1.6.3
